refactor(attention): centralize SM120 prims plan validation
